requirements.txt
pymupdf>=1.23.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
google-generativeai>=0.7.2
streamlit>=1.28.0
